Update README.md
For the system to function properly it will need to be interactive for the drivers, instructors, and client. It must be downloadable for the owner as well. This can be done with contact forms to collect information, an API of map data for pick up and drop off locations, a section for comments and a visual section for the student’s session progress. Also, instructors. The owner will need the application to be cloud based so he can access it from anywhere to update any changes and delete user accounts. This must be done securely so only the owner can view and edit user information. Also, the application must have very minimal technical issues. Furthermore, it is imperative that after the application is up and running, there will be no need to update any code. If so, a very minimal amount.